removed all still existing conversion errors by introducing static_casts
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@270 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/basis/modlinear/algorithm_sweep/LaplaceDownModLinear.hpp
@@ -126,7 +126,7 @@ protected:
             fl += 2.0 * alpha_value;
 		}
 		// right boundary
-		else if(i == (1 << l)-1)
+		else if(static_cast<int>(i) == static_cast<int>((1 << l)-1))
 		{
 			//integration
 			result[seq] = 2.0/3.0 * h * (fl + 2.0*fr)
